de.blizzy.documentr/src/main/java/de/blizzy/documentr/validation/RoleExistsValidator.java
/*
documentr - Edit, maintain, and present software documentation on the web.
Copyright (C) 2012-2013 Maik Schreiber

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <http://www.gnu.org/licenses/>.
*/
package de.blizzy.documentr.validation;

import java.io.IOException;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;
import javax.validation.ValidationException;

import org.apache.commons.lang3.StringUtils;
import org.springframework.beans.factory.annotation.Autowired;

import de.blizzy.documentr.access.UserStore;
import de.blizzy.documentr.validation.annotation.RoleExists;

public class RoleExistsValidator implements ConstraintValidator<RoleExists, String> {
	@Autowired
	private UserStore userStore;

	@Override
	public void initialize(RoleExists annotation) {
	}

	@Override
	public boolean isValid(String value, ConstraintValidatorContext context) {
		if (StringUtils.isBlank(value)) {
			return true;
		}

		try {
			return userStore.getRole(value) != null;
		} catch (IOException e) {
			throw new ValidationException(e);
		}
	}
}
